Split out forms for each journey.
Rather than having a shared form for both journeys use a separate form
for each journey. The attributes shown on the form and persisted to the
eligibility are quite different between each journey so separate forms
make much more sense, hat tip to Giuseppe for the suggestion.
