+++ app/models/journeys/base.rb
@@ -3,7 +3,6 @@ module Journeys
     SHARED_FORMS = {
       "sign-in-or-continue" => SignInOrContinueForm,
       "current-school" => CurrentSchoolForm,
-      "qualification-details" => QualificationDetailsForm,
       "personal-details" => PersonalDetailsForm,
       "select-email" => SelectEmailForm,
       "provide-mobile-number" => ProvideMobileNumberForm
