+++ app/models/journeys/teacher_student_loan_reimbursement.rb
@@ -11,7 +11,8 @@ module Journeys
     POLICIES = [Policies::StudentLoans]
 
     FORMS = {
-      "claim-school" => ClaimSchoolForm
+      "claim-school" => ClaimSchoolForm,
+      "qualification-details" => QualificationDetailsForm
     }
   end
 end
